requirements/base.txt
# Versions are left out on purpose in order for you to use the latest ones.
# After the first "pip install -r", just run "pip freeze" and add the version
# to each package in each requirements/*.txt.

Django==1.10.3
django-rest-swagger==2.0.7
djangorestframework==3.5.3
